+++ packages/cli/src/commands/curl/shared.ts
@@ -181,7 +181,15 @@ export async function getDeploymentUrlAndToken(
     return 1;
   }
 
-  const target = linkedProject.project.latestDeployments?.[0]?.url;
+  /** this is a url like `test-express-5.vercel.app` */
+  const preferredAlias = linkedProject.project.targets?.production?.alias?.[0];
+  /**
+   * this is a url like `test-express-5-yw3u1f2bj-uncurated-tests.vercel.app`
+   *
+   * we're using it as a fallback because as a deployment rolls out there can be a race on getting the `preferredAlias`
+   */
+  const backupAlias = linkedProject.project.latestDeployments?.[0]?.url;
+  const target = preferredAlias || backupAlias;
 
   let baseUrl: string;